Script to upgrade from focal to noble
The script is split into various stages where progress is tracked
on-disk. The script is able to resume where it was at any point, and
needs to, given multiple reboots in the middle.

The new noble-upgrade.json file shipped in the securedrop-config package
is used to control the upgrade process.

Further details of the script are explained inline and at
<https://github.com/freedomofpress/securedrop/wiki/noble-upgrade-architecture>.

Fixes #7332.