Set a fixed APT::Machine-ID so phased updates are consistent
We generally try to keep SecureDrops as homogenus as possible. APT's new
phased updates can cause issues here if different instances are running
different sets of packages.

As explained in <https://discourse.ubuntu.com/t/phased-updates-in-apt-in-21-04/20345>,
we can override the `APT::Machine-ID` configuration setting so that all
instances apply the exact same set of phased updates.

Per machine-id(5), the value is supposed to be a "hexadecimal,
32-character, lowercase ID". The hash is a random value that was
selected; clever reviewers will be able to ascertain the source ;-)

Fixes #7422