Enable manual noble upgrade
Admins can run `./securedrop-admin noble_migration` to trigger a manual
noble migration.

At a high level the playbook:
* disables OSSEC notifications
* triggers the app upgrade, waiting through two reboots
* triggers the mon upgrade, again waiting through reboots
* re-enables OSSEC notifications

The most complicated part is how we for the reboots. We first have a
`wait_for` that looks for a specific stage in the state file. Because
the upgrade script writes the state file and then immediately reboots,
it should never actually succeed and fail because the connection is
interrupted. So we set `ignore_unreachable` and `ignore_errors`, and the
next block is `wait_for_connection` for the server to come back up.
There is a delay before we begin checking just in case the wait_for did
succeed and we need to wait for the reboot to happen.

Because of this sequencing, there isn't any support for the playbook
failing mid-host and restarting it. It is probably unnecessary since,
once started, the upgrade will automatically finish by itself.

The script does support one host already being upgraded and the other
still needing migration. So if e.g. app migration fails, you can
manually fix the host, let it auto finish the upgrade, and then re-run
the playbook to migrate mon.

Fixes #7416.