Speed up pam_ecryptfs testinfra check
After a noble migration, transmitting syslog back to the host so it can
be searched can be pretty slow, so let's do it on the host. By using
grep, we can escape the message we're looking for in a way that ensures
our search doesn't match the error we're looking for.

Since there's already a `.` in the message, use re.escape(), which turns
it into `\.`, which obfuscates our grep.